♻️ refactor : rest doc 테스트 패키지 & 컨텍스트 하나로 리팩토링
