+++ src/test/java/com/coumin/woowahancoupons/coupon/document/CouponRedemptionDocumentationTest.java
@@ -1,4 +1,4 @@
-package com.coumin.woowahancoupons.coupon.controller;
+package com.coumin.woowahancoupons.coupon.document;
 
 import static org.mockito.BDDMockito.given;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
@@ -16,37 +16,20 @@ import static org.springframework.restdocs.request.RequestDocumentation.paramete
 import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import com.coumin.woowahancoupons.ApiDocumentationTest;
 import com.coumin.woowahancoupons.coupon.converter.CouponRedemptionConverter;
 import com.coumin.woowahancoupons.coupon.dto.CouponIssuanceDto;
 import com.coumin.woowahancoupons.coupon.factory.TestCouponFactory;
-import com.coumin.woowahancoupons.coupon.service.CouponRedemptionService;
 import com.coumin.woowahancoupons.domain.coupon.CouponRedemption;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
 import java.util.UUID;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.restdocs.payload.JsonFieldType;
-import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 
-@AutoConfigureRestDocs
-@WebMvcTest(CouponRedemptionRestController.class)
-class CouponRedemptionDocumentationTest {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @MockBean
-    private CouponRedemptionService couponRedemptionService;
-
-    @Autowired
-    private ObjectMapper objectMapper;
+class CouponRedemptionDocumentationTest extends ApiDocumentationTest {
 
     @Test
     @DisplayName("고객이 쿠폰 번호를 입력해 쿠폰 등록")
